feat: Bump version to 0.2.0 and add new app state management and validation features
- Updated package version in package.json to 0.2.0
- Added new app state management in app-state.js
- Implemented input validation for SteamID and API key in app-validators.js
- Created event handlers for SteamID input and API key management in app-events.js
- Enhanced user interface feedback for input validation
- Integrated error handling for vanity URL resolution
